requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
multidict==6.6.1
propcache==0.3.2
python-dotenv==1.1.1
yarl==1.20.1
